% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_functions.R
\name{cran_repo}
\alias{cran_repo}
\alias{cran_list}
\alias{cran_pkg_exists}
\alias{cran_get_version}
\alias{cran_update_check}
\title{CRAN Repo Functions}
\usage{
cran_list()

cran_pkg_exists(pkg_name = NULL)

cran_get_version(pkg_name = NULL)

cran_update_check(pkg_name = NULL)
}
\arguments{
\item{pkg_name}{character of package names}
}
\description{
Functions that interact with cran
}
\section{Functions}{
\itemize{
\item \code{cran_list}: list packages and version numbers. Returns data.table sorted by package, version

\item \code{cran_pkg_exists}: boolean. True if package exists in CRAN

\item \code{cran_get_version}: similar to \code{cran_list} except returns NULL if package does not exist, otherwise
returns the subset of full data returned by \code{cran_list}

\item \code{cran_update_check}: if package does not exist in dsr, returns NULL. Otherwise FALSE if does not need updating, and TRUE
if dsr version is different then CRANs. Will print the respective version numbers to console before returning TRUE.
}}

