% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echo.R
\name{done}
\alias{done}
\title{Prints "DONE!" message}
\usage{
done(toc = TRUE)
}
\arguments{
\item{toc}{If TRUE and `tic()` was used before, displays elapsed time}
}
\description{
Prints "DONE!" message
}
\examples{
done()

}
\author{
Jose Alquicira-Hernandez
}
