% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_functions.R
\name{dsr_repo}
\alias{dsr_repo}
\alias{dsr_list}
\alias{dsr_pkg_exists}
\alias{dsr_get_version}
\alias{dsr_bin_files}
\alias{dsr_src_files}
\alias{dsr_internal_add}
\alias{dsr_internal_rm}
\alias{dsr_update_index}
\alias{dsr_add}
\title{Data Science Repo Functions}
\usage{
dsr_list()

dsr_pkg_exists(pkg_name = NULL)

dsr_get_version(pkg_name = NULL)

dsr_bin_files(pkg_name = NULL)

dsr_src_files(pkg_name = NULL)

dsr_internal_add(pkg_name)

dsr_internal_rm(pkg_name)

dsr_update_index()

dsr_add(pkg_name)
}
\arguments{
\item{pkg_name}{character of package names}
}
\description{
Functions that interact with dsr (repo). Note that these are local functions only, and will run on the R machine
that hosts the repo.
}
\section{Functions}{
\itemize{
\item \code{dsr_list}: list packages and version numbers. Returns data.table sorted by package, version

\item \code{dsr_pkg_exists}: boolean. True if package exists in dsr

\item \code{dsr_get_version}: similar to \code{dsr_list} except returns NULL if package does not exist, otherwise
returns the subset of full data returned by \code{dsr_list}

\item \code{dsr_bin_files}: lists the bin files found in repo

\item \code{dsr_src_files}: lists the source files found in repo

\item \code{dsr_internal_add}: the internal function to add the source and bin package files to the repo

\item \code{dsr_internal_rm}: the internal function to remove the source and bin package files from the repo

\item \code{dsr_update_index}: internal function to update the package index of the repo

\item \code{dsr_add}: user-facing function that adds packages to repo. This will run checks and update if necessary
}}

