% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echo.R
\name{echo}
\alias{echo}
\title{Prints formatted message}
\usage{
echo(string, lb = 2, ellipsis = TRUE)
}
\arguments{
\item{string}{Message to print}

\item{lb}{`2` if break lines are included at the start and end of the string,
`1` final only and `0` for no break lines.}
}
\description{
Prints formatted message
}
\author{
Jose Alquicira-Hernandez
}
