% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echo.R
\name{formatTime}
\alias{formatTime}
\title{Formats time for seconds, minutes and hours}
\usage{
formatTime(x)
}
\arguments{
\item{x}{Numeric value representing seconds}
}
\value{
A list with the format-numeric value and its time unit
}
\description{
Formats time for seconds, minutes and hours
}
\examples{
formatTime(60)
formatTime(60 * 60)

}
\author{
Jose Alquicira-Hernandez
}
