% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance.R
\name{glance}
\alias{glance}
\title{Show a subset from data}
\usage{
glance(x, n = 5)
}
\arguments{
\item{x}{data.frame, matrix, vector, or list object}

\item{n}{index limit}
}
\value{
A subset from provided object. If `n` exceeds dimensions of `x`, returns
maximun dimension
}
\description{
Show a subset from data
}
\examples{
glance(cars)

}
\author{
Jose Alquicira-Hernandez
}
