% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Not in operator}
\usage{
x \%nin\% table
}
\arguments{
\item{x}{Vector or NULL: the values to be matched. Long vectors are supported.}

\item{table}{Vector or NULL: the values to be matched against. Long vectors are not supported.}
}
\description{
nin operator allows using the negation of the in operator
}
\examples{

1:10 \%nin\% 1:5


}
\author{
Jose Alquicira-Hernandez
}
