% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{\%p\%}
\alias{\%p\%}
\title{Concatenates strings via operator}
\usage{
... \%p\% NA
}
\arguments{
\item{...}{strings to paste using paste0()}
}
\description{
Concatenates strings via operator
}
\examples{

"Hello " \%p\% "world"


}
\author{
Jose Alquicira-Hernandez
}
