% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echo.R
\name{inicio}
\alias{inicio}
\title{Prints formatted message}
\usage{
inicio(message = NULL, mem = FALSE)
}
\arguments{
\item{message}{Message to print}

\item{mem}{Track memory usage?}
}
\description{
Prints formatted message
}
\examples{

inicio("Analyzing data")

}
\author{
Jose Alquicira-Hernandez
}
