% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echo.R
\name{init}
\alias{init}
\title{Prints formatted message}
\usage{
init(string, lb = TRUE, ellipsis = TRUE, tic = TRUE)
}
\arguments{
\item{string}{Message to print}

\item{lb}{If `TRUE`, a break line is included at the start of the string}
}
\description{
Prints formatted message
}
\examples{

init("Analyzing data")

}
\author{
Jose Alquicira-Hernandez
}
