% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{print_session}
\alias{print_session}
\title{Prints R session info}
\usage{
print_session(path = NULL, save = TRUE)
}
\arguments{
\item{path}{Path to write file with the session info}

\item{save}{Write session info? If \code{FALSE}, session info is only displayed in console}
}
\description{
Prints R session info
}
\author{
Jose Alquicira-Hernandez
}
