% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{rotate_x}
\alias{rotate_x}
\title{Rotate x-axis labels from ggplot}
\usage{
rotate_x(angle = 90, ...)
}
\arguments{
\item{angle}{rotation angle of x-axis labels}

\item{...}{Additional arguments for theme}
}
\description{
Rotate x-axis labels from ggplot
}
\author{
Jose Alquicira-Hernandez
}
