% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setOutput.R
\name{set_output}
\alias{set_output}
\title{Create output directory}
\usage{
set_output(date, name, sep = "_")
}
\arguments{
\item{date}{Date}

\item{name}{Directory name}

\item{sep}{Character delimiter to separate data from directory name}
}
\description{
Create output directory
}
\examples{
set_output("2019-05-30", "data_analysis")

}
\author{
Jose Alquicira-Hernandez
}
