% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setOutput.R
\name{setOutput}
\alias{setOutput}
\title{Create output directory}
\usage{
setOutput(name, date)
}
\arguments{
\item{name}{Directory name}

\item{date}{Date}
}
\description{
Create output directory
}
\examples{
setOutput("data_analysis", "2019-05-30")

}
\author{
Jose Alquicira-Hernandez
}
