% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_analysis.R
\name{setup_analysis}
\alias{setup_analysis}
\title{Set up analysis template for R script}
\usage{
setup_analysis()
}
\description{
Shiny interface to be used in RStudio to insert a data analysis
template into an empty Rscript document
}
