% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{theme_pub}
\alias{theme_pub}
\title{Theme pub}
\usage{
theme_pub(
  base_size = 14,
  base_family = "",
  border_color = "grey20",
  border_size = 0.6,
  ...
)
}
\arguments{
\item{base_size}{Text size}

\item{base_family}{Text family}

\item{border_color}{Border border color}

\item{border_size}{Border size color}

\item{...}{Additional arguments for theme}
}
\description{
Theme pub
}
\author{
Jose Alquicira-Hernandez
}
