% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echo.R
\name{fin}
\alias{fin}
\title{Prints "DONE!" message}
\usage{
fin(start = "+ ", mem = FALSE)
}
\arguments{
\item{start}{A string symbol}
}
\description{
Prints "DONE!" message
}
\examples{
fin()

}
\author{
Jose Alquicira-Hernandez
}
