% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strings.R
\name{\%p\%}
\alias{\%p\%}
\title{Concatenates strings via operator}
\usage{
"Hello " %p% "world"
}
\arguments{
\item{...}{strings to paste using paste0()}
}
\description{
Concatenates strings via operator
}
\author{
Jose Alquicira-Hernandez
}
