% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echo.R
\name{inicio}
\alias{inicio}
\title{Prints formatted message}
\usage{
inicio(string = NULL, start = "==>", ellipsis = TRUE, mem = TRUE)
}
\arguments{
\item{string}{Message to print}

\item{start}{A string symbol}

\item{ellipsis}{If TRUE, adds ellipsis to message}

\item{mem}{Track memory usage?}
}
\description{
Prints formatted message
}
\examples{

inicio("Analyzing data")

}
\author{
Jose Alquicira-Hernandez
}
