% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setOutput.R
\name{setOutput}
\alias{setOutput}
\title{Create output directory}
\usage{
setOutput(date, name, sep = "_")
}
\arguments{
\item{date}{Date}

\item{name}{Directory name}
}
\description{
Create output directory
}
\examples{
setOutput("2019-05-30", "data_analysis")

}
\author{
Jose Alquicira-Hernandez
}
