% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{theme_lattice}
\alias{theme_lattice}
\title{Theme lattice}
\usage{
theme_lattice(...)
}
\arguments{
\item{...}{Additional arguments for theme}
}
\description{
Theme lattice
}
\author{
Jose Alquicira-Hernandez
}
